Accept slashes in buffer display names
Before this change, if you set the buffer display name to something with
a slash in it, using $PAGER_LABEL or via the embed API, the code would
treat it as a file path and only show the part after the last slash.

With this change in place, requested names are now properly preserved.