Simplify profiling of cold search benchmark
Before this change, we simulated cold searches by recreating the input
between each iteration.

That recreation showed up in profiling, making it hard to see what was
searching and what was cleanup.

With this change in place, we now no longer populate the cache when we
cold benchmark, so no resets necessary.